requirements.txt
Flask
mariadb
Flask-SQLAlchemy
PyMySQL
flask_apscheduler
paho-mqtt
pika
requests